backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<5.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Background Tasks
celery==5.3.6
redis==5.0.1

# WebSockets
websockets==12.0

# Utils
python-dotenv==1.0.0
httpx==0.26.0

# File Processing
python-magic==0.4.27
Pillow==10.2.0

# Testing
pytest==7.4.4
hypothesis==6.98.1
